fix(wizard): keep ExternalLinks mounted; gate only the visible banners
The first cut gated ExternalLinks off the wizard too, which (a) broke
external-links.test.ts's guarantee that the interceptor covers the setup
wizard, and (b) was a real regression — ExternalLinks is an invisible
click-interceptor that makes external <a> links work inside the Tauri
webview, not a visible banner. Hiding it from the wizard gave no visual
benefit but killed the wizard's external links.

Split the gating: ExternalLinks stays in the root layout for every route;
only NoticeBar + NotificationBanner (the visible chrome the issue is about)
are gated off /setup. Rename LayoutChrome → LayoutBanners to match.

Full ui bun test suite (235) green.
