+++ ui/app/layout.tsx
@@ -4,7 +4,8 @@ import type { Metadata } from 'next'
 import { Instrument_Serif, Plus_Jakarta_Sans, JetBrains_Mono } from 'next/font/google'
 import './globals.css'
 import { ThemeProvider } from '@/lib/theme-context'
-import LayoutChrome from '@/components/LayoutChrome'
+import ExternalLinks from '@/components/ExternalLinks'
+import LayoutBanners from '@/components/LayoutBanners'
 
 const instrumentSerif = Instrument_Serif({
   weight: '400',
@@ -45,7 +46,11 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
     >
       <body className="min-h-screen font-sans">
         <ThemeProvider>
-          <LayoutChrome />
+          {/* ExternalLinks is an invisible interceptor — mounted for every
+              route (wizard included) so external links work in the webview.
+              Only the visible banners are gated off the wizard. */}
+          <ExternalLinks />
+          <LayoutBanners />
           {children}
         </ThemeProvider>
       </body>
